+++ apps/platform/pkg/auth/mock/mockauth.go
@@ -1,17 +1,15 @@
 package mock
 
 import (
-	"encoding/json"
 	"errors"
 	"fmt"
-	"log/slog"
 	"net/http"
 
+	"github.com/crewjam/saml"
+	"github.com/crewjam/saml/samlsp"
 	"github.com/thecloudmasters/uesio/pkg/auth"
-	"github.com/thecloudmasters/uesio/pkg/controller/ctlutil"
 	"github.com/thecloudmasters/uesio/pkg/meta"
 	"github.com/thecloudmasters/uesio/pkg/sess"
-	"github.com/thecloudmasters/uesio/pkg/types/exceptions"
 	"github.com/thecloudmasters/uesio/pkg/types/wire"
 )
 
@@ -34,25 +32,15 @@ type Connection struct {
 	session     *sess.Session
 }
 
-func (c *Connection) RequestLogin(w http.ResponseWriter, r *http.Request) {
-	ctlutil.HandleError(r.Context(), w, errors.New("requesting login is not supported by this auth source type"))
-}
-
-func (c *Connection) Login(w http.ResponseWriter, r *http.Request) {
-	var loginRequest map[string]any
-	err := json.NewDecoder(r.Body).Decode(&loginRequest)
+func (c *Connection) Login(loginRequest auth.AuthRequest) (*auth.LoginResult, error) {
+	user, loginMethod, err := c.DoLogin(loginRequest)
 	if err != nil {
-		const msg = "invalid login request body"
-		slog.InfoContext(r.Context(), fmt.Sprintf("%s: %v", msg, err))
-		ctlutil.HandleError(r.Context(), w, exceptions.NewBadRequestException(msg, nil))
-		return
+		return nil, err
 	}
-	user, _, err := c.DoLogin(loginRequest)
-	if err != nil {
-		ctlutil.HandleError(r.Context(), w, err)
-		return
-	}
-	auth.LoginRedirectResponse(w, r, user, c.session)
+	return &auth.LoginResult{
+		AuthResult:    auth.AuthResult{User: user, LoginMethod: loginMethod},
+		PasswordReset: false,
+	}, nil
 }
 
 func (c *Connection) DoLogin(payload map[string]any) (*meta.User, *meta.LoginMethod, error) {
@@ -76,4 +64,10 @@ func (c *Connection) CreateLogin(signupMethod *meta.SignupMethod, payload map[st
 }
 func (c *Connection) ConfirmSignUp(signupMethod *meta.SignupMethod, payload map[string]any) error {
 	return errors.New("mock login: unfortunately you cannot change the password")
+}
+func (c *Connection) GetServiceProvider(r *http.Request) (*samlsp.Middleware, error) {
+	return nil, errors.New("saml auth is not supported by this auth source type")
+}
+func (c *Connection) LoginServiceProvider(assertion *saml.Assertion) (*auth.LoginResult, error) {
+	return nil, errors.New("saml auth login is not supported by this auth source type")
 }
